+++ src/arbitrary/_std/time.rs
@@ -10,9 +10,11 @@
 //! Arbitrary implementations for `std::time`.
 
 use std::time::*;
+use core::ops::Range;
 
-use strategy::statics::static_map;
+use strategy::statics::{self, static_map};
 use arbitrary::*;
+use num;
 
 arbitrary!(Duration, SMapped<(u64, u32), Self>;
     static_map(any::<(u64, u32)>(), |(a, b)| Duration::new(a, b))
@@ -21,16 +23,24 @@ arbitrary!(Duration, SMapped<(u64, u32), Self>;
 // Instant::now() "never" returns the same Instant, so no shrinking may occur!
 arbitrary!(Instant; Self::now());
 
-// Same for SystemTime.
-arbitrary!(SystemTime; Self::now());
-
-/*
-A possible logic for SystemTimeError:
-fn gen_ste() -> SystemTimeError {
-    (SystemTime::now() + Duration::from_millis(10)).elapsed().unwrap_err()
-}
-This may however panic from time to time. NTP could also ruin our day!
-*/
+arbitrary!(
+    // We can't use `any::<Duration>()` because the addition to `SystemTime`
+    // can overflow and panic. To be conservative, we only allow seconds to go
+    // to i32::MAX since a certain popular OS still uses `i32` to represent the
+    // seconds counter.
+    SystemTime, statics::Map<(num::i32::Any, Range<u32>),
+                             fn ((i32, u32)) -> SystemTime>;
+    static_map((num::i32::ANY, 0..1_000_000_000u32),
+                |(sec, ns)| {
+                    if sec >= 0 {
+                        SystemTime::UNIX_EPOCH + Duration::new(
+                            sec as u64, ns)
+                    } else {
+                        SystemTime::UNIX_EPOCH - Duration::new(
+                            (-(sec as i64)) as u64, ns)
+                    }
+                })
+);
 
 #[cfg(test)]
 mod test {