+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ### New Additions
 
-- Added `max_shrink_iters` and `max_shrink_time` options to test configuration
-  to allow capping the resources expended on shrinking test cases.
+- Add `max_shrink_iters` and `max_shrink_time` options to test configuration to
+  allow capping the resources expended on shrinking test cases.
+
+- Add `verbose` option to make proptest give details about what is happening as
+  the test executes.
 
 ### Bug Fixes